Moved filter.py to datafilereaders folder
